Add ability to parse an eml file into a FileMessage
