+++ src/Message.php
@@ -3,25 +3,15 @@
 namespace DirectoryTree\ImapEngine;
 
 use BackedEnum;
-use Carbon\Carbon;
 use DirectoryTree\ImapEngine\Enums\ImapFlag;
-use DirectoryTree\ImapEngine\Exceptions\RuntimeException;
 use DirectoryTree\ImapEngine\Support\Str;
 use Illuminate\Contracts\Support\Arrayable;
 use JsonSerializable;
 use Stringable;
-use ZBateson\MailMimeParser\Header\HeaderConsts;
-use ZBateson\MailMimeParser\Header\IHeader;
-use ZBateson\MailMimeParser\Header\Part\AddressPart;
-use ZBateson\MailMimeParser\Message as MailMimeMessage;
-use ZBateson\MailMimeParser\Message\MimePart;
 
 class Message implements Arrayable, JsonSerializable, Stringable
 {
-    /**
-     * The parsed message.
-     */
-    protected ?MailMimeMessage $parsed = null;
+    use HasParsedMessage;
 
     /**
      * Constructor.
@@ -92,156 +82,6 @@ class Message implements Arrayable, JsonSerializable, Stringable
     }
 
     /**
-     * Get the message date and time.
-     */
-    public function date(): ?Carbon
-    {
-        if ($date = $this->header(HeaderConsts::DATE)?->getDateTime()) {
-            return Carbon::instance($date);
-        }
-
-        return null;
-    }
-
-    /**
-     * Get the message's message-id.
-     */
-    public function messageId(): ?string
-    {
-        return $this->header(HeaderConsts::MESSAGE_ID)?->getValue();
-    }
-
-    /**
-     * Get the message's subject.
-     */
-    public function subject(): ?string
-    {
-        return $this->header(HeaderConsts::SUBJECT)?->getValue();
-    }
-
-    /**
-     * Get the FROM address.
-     */
-    public function from(): ?Address
-    {
-        return head($this->addresses(HeaderConsts::FROM)) ?: null;
-    }
-
-    /**
-     * Get the SENDER address.
-     */
-    public function sender(): ?Address
-    {
-        return head($this->addresses(HeaderConsts::SENDER)) ?: null;
-    }
-
-    /**
-     * Get the REPLY-TO address.
-     */
-    public function replyTo(): ?Address
-    {
-        return head($this->addresses(HeaderConsts::REPLY_TO)) ?: null;
-    }
-
-    /**
-     * Get the IN-REPLY-TO address.
-     */
-    public function inReplyTo(): ?Address
-    {
-        return head($this->addresses(HeaderConsts::IN_REPLY_TO)) ?: null;
-    }
-
-    /**
-     * Get the TO addresses.
-     *
-     * @return Address[]
-     */
-    public function to(): array
-    {
-        return $this->addresses(HeaderConsts::TO);
-    }
-
-    /**
-     * Get the CC addresses.
-     *
-     * @return Address[]
-     */
-    public function cc(): array
-    {
-        return $this->addresses(HeaderConsts::CC);
-    }
-
-    /**
-     * Get the BCC addresses.
-     *
-     * @return Address[]
-     */
-    public function bcc(): array
-    {
-        return $this->addresses(HeaderConsts::BCC);
-    }
-
-    /**
-     * Get the message's attachments.
-     *
-     * @return Attachment[]
-     */
-    public function attachments(): array
-    {
-        return array_map(function (MimePart $part) {
-            return new Attachment(
-                $part->getFilename(),
-                $part->getContentType(),
-                $part->getContentStream(),
-            );
-        }, $this->parse()->getAllAttachmentParts());
-    }
-
-    /**
-     * Determine if the message has attachments.
-     */
-    public function hasAttachments(): bool
-    {
-        return $this->attachmentCount() > 0;
-    }
-
-    /**
-     * Get the count of attachments.
-     */
-    public function attachmentCount(): int
-    {
-        return $this->parse()->getAttachmentCount();
-    }
-
-    /**
-     * Get addresses from the given header.
-     *
-     * @return Address[]
-     */
-    public function addresses(string $header): array
-    {
-        return array_map(function (AddressPart $part) {
-            return new Address($part->getEmail(), $part->getName());
-        }, $this->header($header)?->getParts() ?? []);
-    }
-
-    /**
-     * Get the message's HTML content.
-     */
-    public function html(): ?string
-    {
-        return $this->parse()->getHtmlContent();
-    }
-
-    /**
-     * Get the message's text content.
-     */
-    public function text(): ?string
-    {
-        return $this->parse()->getTextContent();
-    }
-
-    /**
      * Determine if the message is marked as seen.
      */
     public function isSeen(): bool
@@ -295,22 +135,6 @@ class Message implements Arrayable, JsonSerializable, Stringable
     public function hasFlag(BackedEnum|string $flag): bool
     {
         return in_array(Str::enum($flag), $this->flags);
-    }
-
-    /**
-     * Get all headers from the message.
-     */
-    public function headers(): array
-    {
-        return $this->parse()->getAllHeaders();
-    }
-
-    /**
-     * Get a header from the message.
-     */
-    public function header(string $name, int $offset = 0): ?IHeader
-    {
-        return $this->parse()->getHeader($name, $offset);
     }
 
     /**
@@ -487,18 +311,6 @@ class Message implements Arrayable, JsonSerializable, Stringable
     }
 
     /**
-     * Parse the message into a MailMimeMessage instance.
-     */
-    public function parse(): MailMimeMessage
-    {
-        if (! $this->hasHead() && ! $this->hasBody()) {
-            throw new RuntimeException('Cannot parse an empty message');
-        }
-
-        return $this->parsed ??= MessageParser::parse((string) $this);
-    }
-
-    /**
      * Get the array representation of the message.
      */
     public function toArray(): array
@@ -528,5 +340,13 @@ class Message implements Arrayable, JsonSerializable, Stringable
     public function jsonSerialize(): array
     {
         return $this->toArray();
+    }
+
+    /**
+     * Determine if the message is empty.
+     */
+    protected function isEmpty(): bool
+    {
+        return ! $this->hasHead() && ! $this->hasBody();
     }
 }
